Mesto ПР7

+++ src/components/card.js
@@ -1,29 +1,70 @@
-
+import { showError, removeCard, likeCard, unlikeCard } from './api.js';
 // @todo: Темплейт карточки
-const cardTemplate = document.querySelector("#card-template").content;
+const cardTemplate = document.querySelector('#card-template').content;
 
 // @todo: Функция создания карточки
-function createCard(cardItem, deleteCard, addLike, clickImage) {
-  const cardElement = cardTemplate.querySelector(".card").cloneNode(true);
-  const cardImage = cardElement.querySelector(".card__image");
+function createCard(cardItem, userId, deleteCard, clickImage) {
+  const cardElement = cardTemplate.querySelector('.card').cloneNode(true);
+  const cardImage = cardElement.querySelector('.card__image');
+  const cardLike = cardElement.querySelector('.card__like-count');
+  const cardDelButton = cardElement.querySelector('.card__delete-button');
 
-  cardElement.querySelector(".card__title").textContent = cardItem.name;
+  cardElement.querySelector('.card__title').textContent = cardItem.name;
   cardImage.src = cardItem.link;
   cardImage.alt = cardItem.name;
 
+  cardLike.textContent = cardItem.likes.length;
+
+  // Скрыть корзину на чужих карточках
+  if (cardItem.owner._id !== userId) {
+    cardDelButton.style.display = 'none';
+  }
+
   //Обработчик удаления карточки
-  const buttonDel = cardElement.querySelector(".card__delete-button");
-  buttonDel.addEventListener("click", function () {
-    deleteCard(cardElement);
+  const buttonDel = cardElement.querySelector('.card__delete-button');
+  if (cardItem.owner._id === userId)
+    buttonDel.addEventListener('click', () => {
+      removeCard(cardItem._id)
+        .then(() => {
+          deleteCard(cardElement);
+        })
+        .catch(showError);
+    });
+
+  const buttonLike = cardElement.querySelector('.card__like-button');
+
+  // Проверяем, лайкнули ли мы карточку
+  const likedByMe = cardItem.likes.some((like) => like._id === userId);
+  if (likedByMe) {
+    buttonLike.classList.add('card__like-button_is-active');
+  }
+  // Обработчик лайка
+  buttonLike.addEventListener('click', () => {
+    const activeLike = buttonLike.classList.contains(
+      'card__like-button_is-active'
+    );
+
+    if (activeLike) {
+      unlikeCard(cardItem._id)
+        .then((updatedCard) => {
+          buttonLike.classList.remove('card__like-button_is-active');
+          cardLike.textContent = updatedCard.likes.length;
+          cardItem.likes = updatedCard.likes;
+        })
+        .catch(showError);
+    } else {
+      likeCard(cardItem._id)
+        .then((updatedCard) => {
+          buttonLike.classList.add('card__like-button_is-active');
+          cardLike.textContent = updatedCard.likes.length;
+          cardItem.likes = updatedCard.likes;
+        })
+        .catch(showError);
+    }
   });
 
-  //обработчик лайка
-  const buttonLike = cardElement.querySelector(".card__like-button");
-  buttonLike.addEventListener("click", addLike);
-
   //обработчик клика по изображению
-  cardImage.addEventListener("click", clickImage);
-
+  cardImage.addEventListener('click', clickImage);
   return cardElement;
 }
 
@@ -32,9 +73,4 @@ function deleteCard(cardElement) {
   cardElement.remove();
 }
 
-// Функция добавления лайка
-function addLike(evt) {
-  evt.target.classList.toggle("card__like-button_is-active");
-}
-
-export { createCard, deleteCard, addLike };
+export { createCard, deleteCard };
